backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
firebase-admin==6.2.0
pymongo==4.6.0 